+++ metadata_fetcher/fetchers/OACFetcher.py
@@ -1,7 +1,7 @@
 import json
 import requests
 from xml.etree import ElementTree
-from Fetcher import Fetcher
+from .Fetcher import Fetcher
 
 
 class OACFetcher(Fetcher):
